core/src/test/java/ch/cyberduck/core/diagnostics/DiskReachabilityTest.java
package ch.cyberduck.core.diagnostics;

/*
 * Copyright (c) 2002-2023 iterate GmbH. All rights reserved.
 * https://cyberduck.io/
 *
 * This program is free software; you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 */

import ch.cyberduck.core.Host;
import ch.cyberduck.core.Scheme;
import ch.cyberduck.core.TestProtocol;

import org.junit.Test;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

public class DiskReachabilityTest {

    @Test
    public void isReachable() {
        final Host bookmark = new Host(new TestProtocol(Scheme.file));
        assertTrue(new DiskReachability().isReachable(bookmark));
        bookmark.setDefaultPath("/notfound");
        assertFalse(new DiskReachability().isReachable(bookmark));
    }
}